Consolidate names around RNTrackPlayer
Project was getting detected as both RNTrackPlayer and TrackPlayer.
